scripts/requirements.txt
webuiapi
pillow
requests
torch
